Single all inventory page
This addition marks the completion of the single all inventory page issue for the 2.0 project. In addition, it cleans up the Category.tsx component with the addition of an ItemImages.ts that takes the item image assignments out of the Category.tsx file making it easier to work with for any future changes.